Conflicts:
	deps/hiredis/async.c
	deps/hiredis/examples/example-ae.c
	deps/hiredis/hiredis.c
	deps/hiredis/hiredis.h
	deps/hiredis/net.c
	deps/hiredis/net.h
	deps/jemalloc/bin/pprof
	msvs/RedisServer.vcxproj
	src/Win32_Interop/Win32_CommandLine.cpp
	src/Win32_Interop/Win32_QFork.cpp
	src/Win32_Interop/Win32_QFork.h
	src/Win32_Interop/win32_wsiocp.c
	src/Win32_Interop/win32fixes.c
	src/anet.c
	src/config.c
	src/networking.c
	src/pubsub.c
	src/rand.c
	src/rdb.c
	src/redis-benchmark.c
	src/redis-cli.c
	src/redis.c
	src/redis.h
	src/zmalloc.c
	tests/integration/replication-psync.tcl
	tests/unit/aofrw.tcl

+++ src/replication.c
@@ -197,10 +197,10 @@ void replicationFeedSlaves(list *slaves, int dictid, robj **argv, int argc) {
         feedReplicationBacklog(aux,len+3);
 
         for (j = 0; j < argc; j++) {
-            long objlen = stringObjectLen(argv[j]);
+            long objlen = (long)stringObjectLen(argv[j]);
 
             /* We need to feed the buffer with the object as a bulk reply
-             * not just as a plain string, so create the $..CRLF payload len 
+             * not just as a plain string, so create the $..CRLF payload len
              * ad add the final CRLF */
             aux[0] = '$';
             len = ll2string(aux+1,sizeof(aux)-1,objlen);
@@ -2379,7 +2379,7 @@ void replicationCron(void) {
     if (server.masterhost && server.master &&
         !(server.master->flags & REDIS_PRE_PSYNC))
         replicationSendAck();
-    
+
     /* If we have attached slaves, PING them from time to time.
      * So slaves can implement an explicit timeout to masters, and will
      * be able to detect a link disconnection even if the TCP connection
